move create and exercise endpoints to separate module
